% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{package}
\name{fable.timegpt-package}
\alias{fable.timegpt}
\alias{fable.timegpt-package}
\title{fable.timegpt: TimeGPT Modelling Interface for 'fable'}
\description{
Allows TimeGPT forecasts from the Nixtla API to be used in a tidy workflow with the modelling interface of 'fabletools'. This extends TimeGPT to provide enhanced model specification and management, performance evaluation methods, and model combination tools.
}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

}
\keyword{package}
