% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{forecast.fbl_timegpt}
\alias{forecast.fbl_timegpt}
\title{Produce forecasts from the TimeGPT API}
\usage{
\method{forecast}{fbl_timegpt}(
  object,
  new_data,
  specials = NULL,
  times = 1000,
  level = NULL,
  token = Sys.getenv("TIMEGPT_TOKEN"),
  ...
)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{times}{The number of sample paths to use in estimating the forecast distribution when \code{bootstrap = TRUE}.}

\item{level}{The level(s) of uncertainty to produce conformal forecast intervals for.}

\item{token}{The API token obtained from \url{https://dashboard.nixtla.io/}}

\item{...}{Unused.}
}
\value{
A fable.
}
\description{
Produce forecasts from the TimeGPT API
}
