% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{TimeGPT}
\alias{TimeGPT}
\title{TimeGPT SDK for generative forecasting}
\usage{
TimeGPT(formula, ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted of class \code{formula}.}

\item{...}{Additional arguments passed to the API (see \url{https://docs.nixtla.io/reference/timegpt_timegpt_post})}
}
\description{
Provides a fable compatible model to obtain forecasts using the Nixtla's
TimeGPT API.
}
\details{
The TimePGT modelling interface uses a \code{formula} based model specification
(\code{y ~ x}), where the left of the formula specifies the response variable,
and the right specifies the model's predictive terms. Like any model in the
fable framework, it is possible to specify transformations on the response.

The introduction vignette provides more details on how to model data
using this interface to the TimeGPT API: \code{vignette("intro", package="fable.timegpt")}.
}
\section{Specials}{


There are no specials for this forecasting model.
}

\examples{
library(tsibble)
as_tsibble(USAccDeaths) \%>\%
  model(
    timegpt = TimeGPT(value)
  )

}
\seealso{
\itemize{
\item \href{https://docs.nixtla.io/}{TimeGPT homepage}
\item \href{https://docs.nixtla.io/docs}{TimeGPT documentation}
\item \href{https://docs.nixtla.io/docs/python-sdk-installation}{TimeGPT Python SDK}
}
}
